refactor: move game scene manager to managers package
